--- src/redux/categories.js
@@ -1,14 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = [];
-
-const categoriesSlice = createSlice({
-  name: 'categories',
-  initialState,
-  reducers: {
-    checkStatus: () => 'Under Construction',
-  },
-});
-
-export const categoriesAction = categoriesSlice.actions;
-export default categoriesSlice.reducer;
