+++ src/redux/categories/categories.js
@@ -1,0 +1,21 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialState = [];
+
+export const CHECKSTATUS = 'check-status';
+
+// Action Creators
+export function checkStatus() {
+  return { type: CHECKSTATUS };
+}
+
+// Reducer
+export default (state = initialState, action) => {
+  switch (action.type) {
+    case CHECKSTATUS: {
+      return 'Under construction';
+    }
+    default:
+      return state;
+  }
+};
