+++ src/redux/configureStore.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
-import books from './books';
-import categories from './categories';
+import books from './books/books';
+import categories from './categories/categories';
 
 const store = configureStore({
   reducer: {
